Move continue class code to separate file
