+++ python/lsst/sphgeom/__init__.py
@@ -28,8 +28,8 @@
 
 """lsst.sphgeom
 """
-import typing
 
+from ._continue_class import *
 from ._healpixPixelization import *
 from ._sphgeom import *
 from ._sphgeom import Pixelization
@@ -38,130 +38,3 @@ from .pixelization_abc import *
 from .version import *
 
 PixelizationABC.register(Pixelization)
-
-# Copy and paste from lsst.utils.wrappers
-_INTRINSIC_SPECIAL_ATTRIBUTES = frozenset(
-    (
-        "__qualname__",
-        "__module__",
-        "__metaclass__",
-        "__dict__",
-        "__weakref__",
-        "__class__",
-        "__subclasshook__",
-        "__name__",
-        "__doc__",
-    )
-)
-
-
-def _isAttributeSafeToTransfer(name: str, value: typing.Any) -> bool:
-    if name.startswith("__") and (
-        value is getattr(object, name, None) or name in _INTRINSIC_SPECIAL_ATTRIBUTES
-    ):
-        return False
-    return True
-
-
-def _continueClass(cls):
-    import sys
-
-    orig = getattr(sys.modules[cls.__module__], cls.__name__)
-    for name in dir(cls):
-        # Common descriptors like classmethod and staticmethod can only be
-        # accessed without invoking their magic if we use __dict__; if we use
-        # getattr on those we'll get e.g. a bound method instance on the dummy
-        # class rather than a classmethod instance we can put on the target
-        # class.
-        attr = cls.__dict__.get(name, None) or getattr(cls, name)
-        if _isAttributeSafeToTransfer(name, attr):
-            setattr(orig, name, attr)
-    return orig
-
-
-def _inf_to_lat(lat: float) -> float:
-    """Map +Inf to +90 and -Inf to -90 degrees."""
-    import math
-
-    if not math.isinf(lat):
-        return lat
-    if lat > 0.0:
-        return 90.0
-    return -90.0
-
-
-@_continueClass
-class Region:
-    """A minimal interface for 2-dimensional regions on the unit sphere."""
-
-    @classmethod
-    def from_ivoa_pos(cls, pos: str) -> Region:
-        """Create a Region from an IVOA POS string.
-
-        Parameters
-        ----------
-        pos : `str`
-            A string using the IVOA SIAv2 POS syntax.
-
-        Returns
-        -------
-        region : `Region`
-            A region equivalent to the POS string.
-
-        Notes
-        -----
-        See
-        https://ivoa.net/documents/SIA/20151223/REC-SIA-2.0-20151223.html#toc12
-        for a description of the POS parameter but in summary the options are:
-
-        * ``CIRCLE <longitude> <latitude> <radius>``
-        * ``RANGE <longitude1> <longitude2> <latitude1> <latitude2>``
-        * ``POLYGON <longitude1> <latitude1> ... (at least 3 pairs)``
-
-        Units are degrees in all coordinates.
-        """
-        shape, *coordinates = pos.split()
-        coordinates = tuple(float(c) for c in coordinates)
-        n_floats = len(coordinates)
-        if shape == "CIRCLE":
-            if n_floats != 3:
-                raise ValueError(f"CIRCLE requires 3 numbers but got {n_floats} in '{pos}'.")
-            center = LonLat.fromDegrees(coordinates[0], coordinates[1])
-            radius = Angle.fromDegrees(coordinates[2])
-            return Circle(UnitVector3d(center), radius)
-        elif shape == "RANGE":
-            import math
-
-            if n_floats != 4:
-                raise ValueError(f"RANGE requires 4 numbers but got {n_floats} in '{pos}'.")
-            # POS allows +Inf and -Inf in ranges. These are not allowed by
-            # Box and so must be converted.
-            # If either of the longitude values are infinite, all longitudes
-            # should be included.
-            if math.isinf(coordinates[0]) or math.isinf(coordinates[1]):
-                return Box(
-                    NormalizedAngleInterval.fromDegrees(0.0, 360.0),
-                    AngleInterval.fromDegrees(_inf_to_lat(coordinates[2]), _inf_to_lat(coordinates[3])),
-                )
-            else:
-                return Box(
-                    LonLat.fromDegrees(coordinates[0], _inf_to_lat(coordinates[2])),
-                    LonLat.fromDegrees(coordinates[1], _inf_to_lat(coordinates[3])),
-                )
-        elif shape == "POLYGON":
-            if n_floats % 2 != 0:
-                raise ValueError(f"POLYGON requires even number of floats but got {n_floats} in '{pos}'.")
-            if n_floats < 6:
-                raise ValueError(
-                    f"POLYGON specification requires at least 3 coordinates, got {n_floats // 2} in '{pos}'"
-                )
-            # Coordinates are x1, y1, x2, y2, x3, y3...
-            # Get pairs by skipping every other value.
-            pairs = list(zip(coordinates[0::2], coordinates[1::2]))
-            vertices = [LonLat.fromDegrees(lon, lat) for lon, lat in pairs]
-            return ConvexPolygon([UnitVector3d(c) for c in vertices])
-
-        raise ValueError(f"Unrecognized shape in POS string '{pos}'")
-
-
-del typing
